pyproject.toml
[tool.poetry]
name = "carbon-activity"
version = "0.1.0"
description = "Calculate Carbon Activity for a Climate Conscious Future"
authors = ["pandeakshat"]
license = "MT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
pillow = "^10.4.0"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
